Fixed duplicate code and more cleanup
